### 2023-03-06 90e5c4e
removing deep copies of air temps for RH correction
