### 2023-03-06 b765fce
Removing deep-copied T_u from CloudCoverage
