### 2023-03-06 a7b83df
DOIs replaced with thredds links
