+++ .github/workflows/url_update.yml
@@ -1,6 +1,4 @@
 on:
-  schedule:
-    - cron: "0 12 1 * *"
   workflow_dispatch:
   
 jobs:
