+++ src/pypromice/get.py
@@ -3,26 +3,41 @@
 """
 AWS data retrieval module
 """
+from pyDataverse.api import NativeApi
 import pandas as pd
+import xarray as xr
 import unittest, pkg_resources
 from datetime import datetime
         
-def aws_names(url_index=None, delimiter=','):
+
+def lookup_table(base_dois,
+                 server='https://dataverse.geus.dk'):
+    # Prime API
+    dataverse_server = server.strip("/")
+    api = NativeApi(dataverse_server)
+    
+    # Look through DOI entries
+    lookup_list = {}
+    for d in base_dois:
+        dataset = api.get_dataset(d)
+        
+        # Get file names and DOIs
+        f_list = dataset.json()['data']['latestVersion']['files']
+        for f in f_list:
+            fname = f['dataFile']['filename'].lower()
+            if '.csv' in fname or '.nc' in fname or '.txt' in fname:
+                link = _getURL(f['dataFile']['persistentId'])
+                lookup_list[fname] = link
+    return lookup_list       
+   
+def aws_names():
     '''Return PROMICE and GC-Net AWS names that can be used in get.aws_data() 
     fetching'''
-    if url_index is None:
-        with pkg_resources.resource_stream('pypromice', 'data_urls.csv') as stream:
-            lines = stream.read().decode("utf-8")
-            lines = lines.split("\n")
-    else:
-        stream = url_index 
-        with open(stream, 'r') as f:
-            lines = f.readlines()
-    names = [l.split(delimiter)[0] for l in lines]
-    print(f'Available dataset keywords: {names}')
-    return names    
-    
-def aws_data(aws_name):                                                        #TODO add daily and monthly datasets
+    lookup = lookup_table(['doi:10.22008/FK2/IW73UU', 'doi:10.22008/FK2/GNYFUK'])
+    print(f'Available dataset keywords: {list(lookup.keys())}')
+    return list(lookup.keys())
+
+def aws_data(aws_name):
     '''Return PROMICE and GC-Net AWS L3 v3 hourly observations
     
     Returns
@@ -30,135 +45,51 @@ def aws_data(aws_name):                                                        #
     df : pandas.DataFrame
         AWS observations dataframe
     '''
-    URL = _getURL(aws_name.lower()+'_hour')
-    df = pd.read_csv(URL, index_col=0, parse_dates=True)
-    return df
-    
-def watson_discharge_hourly():
+    lookup = lookup_table(['doi:10.22008/FK2/IW73UU', 'doi:10.22008/FK2/GNYFUK'])
+    assert aws_name.lower() in list(lookup.keys())
+    if '.csv' in aws_name.lower():
+        data = pd.read_csv(lookup[aws_name], index_col=0, parse_dates=True)
+    elif '.nc' in aws_name.lower():
+        data = xr.open_dataset(lookup[aws_name])
+    return data        
+  
+def watson_discharge(t='hour'):
     '''Return PROMICE hourly Watson river discharge
     
+    Parameters
+    ----------
+    t : str
+        Temporal resolution of the data - "hour", "day" or "month"
+
     Returns
     -------
     df : pandas.DataFrame
         Watson river discharge dataframe    
     '''
-    URL = _getURL('watson_discharge_hourly')
-    df = pd.read_csv(URL, sep="\s+", parse_dates=[[0,1,2,3]])\
-            .rename({"WaterFluxDiversOnly(m3/s)"         : "divers",
-                    "Uncertainty(m3/s)"                 : "divers_err",
-                    "WaterFluxDivers&Temperature(m3/s)" : "divers_t",
[…]
