### 2023-03-07 78cbfc8
look-up table implemented in module
