+++ src/pypromice/get.py
@@ -8,27 +8,6 @@ import pandas as pd
 import xarray as xr
 import unittest, pkg_resources
 from datetime import datetime
-        
-
-def lookup_table(base_dois,
-                 server='https://dataverse.geus.dk'):
-    # Prime API
-    dataverse_server = server.strip("/")
-    api = NativeApi(dataverse_server)
-    
-    # Look through DOI entries
-    lookup_list = {}
-    for d in base_dois:
-        dataset = api.get_dataset(d)
-        
-        # Get file names and DOIs
-        f_list = dataset.json()['data']['latestVersion']['files']
-        for f in f_list:
-            fname = f['dataFile']['filename'].lower()
-            if '.csv' in fname or '.nc' in fname or '.txt' in fname:
-                link = _getURL(f['dataFile']['persistentId'])
-                lookup_list[fname] = link
-    return lookup_list       
    
 def aws_names():
     '''Return PROMICE and GC-Net AWS names that can be used in get.aws_data() 
@@ -47,10 +26,7 @@ def aws_data(aws_name):
     '''
     lookup = lookup_table(['doi:10.22008/FK2/IW73UU', 'doi:10.22008/FK2/GNYFUK'])
     assert aws_name.lower() in list(lookup.keys())
-    if '.csv' in aws_name.lower():
-        data = pd.read_csv(lookup[aws_name], index_col=0, parse_dates=True)
-    elif '.nc' in aws_name.lower():
-        data = xr.open_dataset(lookup[aws_name])
+    data = pd.read_csv(lookup[aws_name], index_col=0, parse_dates=True)
     return data        
   
 def watson_discharge(t='hour'):
@@ -59,7 +35,7 @@ def watson_discharge(t='hour'):
     Parameters
     ----------
     t : str
-        Temporal resolution of the data - "hour", "day" or "month"
+        Temporal resolution of the data - "hour", "day" or "year"
 
     Returns
     -------
@@ -67,28 +43,74 @@ def watson_discharge(t='hour'):
         Watson river discharge dataframe    
     '''
     lookup = lookup_table(['doi:10.22008/FK2/XEHYCM'])
-    if 'month' in t.lower():
-        link = lookup['watson river discharge (2006-2021) monthly.txt']
+    
+    if 'year' in t.lower():
+        link = lookup['watson river discharge (1949-2021) yearly.txt']
+        df = pd.read_csv(link, sep="\s+", skiprows=9, index_col=0)
+        
     elif 'daily' in t.lower() or 'day' in t.lower():
-        link = lookup['watson river discharge (2006-2021) daily.txt']     
+        link = lookup['watson river discharge (2006-2021) daily.txt']   
+        df = pd.read_csv(link, sep="\s+", parse_dates=[[0,1,2]])\
+                        .rename({"WaterFluxDiversOnly(m3/s)"        : "divers",
+                                "Uncertainty(m3/s)"                 : "divers_err",
+                                "WaterFluxDivers&Temperature(m3/s)" : "divers_t",
+                                "Uncertainty(m3/s).1"               : "divers_t_err",
+                                "WaterFluxCumulative(km3)"          : "Q",
+                                "Uncertainty(km3)"                  : "err"}, 
+                                axis='columns')
+        df['time'] = df.iloc[:,0]
+        df = df.set_index('time')
+        df.drop(columns=df.columns[0:1], axis=1, inplace=True) 
+        
     else:
         link = lookup['watson river discharge (2006-2021) hourly.txt']  
         
-    df = pd.read_csv(link, sep="\s+", parse_dates=[[0,1,2,3]])\
-                    .rename({"WaterFluxDiversOnly(m3/s)"         : "divers",
[…]
