+++ paper/paper.bib
@@ -4,6 +4,7 @@
     title = {A new programme for monitoring the mass loss of the Greenland ice sheet},
     year = {2008},
     url = {https://doi.org/10.34194/geusb.v15.5045},
+    doi = {10.34194/geusb.v15.5045},
     volume = {15},	
     pages = {61-64},
     journal = {GEUS Bulletin},
@@ -13,6 +14,7 @@
     author = {Box, J.E. and Hubbard, A. and Bahr, D. B. and Colgan, W. T. and Fettweis, X. and Mankoff, K. D. and Wehrlé, A. and Noël, B. and van den Broeke, M. R. and Wouters, B. and Bjørk, A. A. and Fausto R. S.},
     title = {Greenland ice sheet climate disequilibrium and committed sea-level rise},
     year = {2022},
+    doi = {10.1038/s41558-022-01441-2},
     url = {https://doi.org/10.1038/s41558-022-01441-2},
     volume = {12},	
     pages = {808-813},
@@ -106,6 +108,7 @@ Publisher: Ubiquity Press},
     author = {Machguth, H. and MacFerrin, M. and van As, D. and Box, J. E. and Charalampidis, C. and Colgan, W. and Fausto, R. S. and Meijer, H. A. J. and Mosley-Thompson, E. and van de Wal, R. S. W.},
     title = {Greenland meltwater storage in firn limited by near-surface ice formation},
     year = {2016},
+    doi = {10.1038/nclimate2899},
     url = {https://doi.org/10.1038/nclimate2899},
     volume = {6},	
     pages = {390-393},
@@ -117,6 +120,7 @@ Publisher: Ubiquity Press},
 	volume = {10},
 	issn = {2296-6463},
 	url = {https://doi.org/10.3389/feart.2022.970026},
+	doi = {10.3389/feart.2022.970026},
 	urldate = {2022-12-15},
 	journal = {Frontiers in Earth Science},
 	author = {Messerli, Alexandra and Arthur, Jennifer and Langley, Kirsty and How, Penelope and Abermann, Jakob},
@@ -126,6 +130,7 @@ Publisher: Ubiquity Press},
 @article{moon-greenland-2022a,
     author = {Moon, T. A. and Mankoff, K. D. and Fausto, R. S. and Fettweis, X. and Loomis, B. D. and Mote, T. L. and Poinar, K. and Tedesco, M. and Wehrlé, A. and Jensen, C. D.},
     title = {Greenland ice sheet},
+    doi = {10.25923/c430-hb50},
     year = {2022},
     url = {https://doi.org/10.25923/c430-hb50},
     journal = {Arctic report card 2022},
@@ -136,6 +141,7 @@ Publisher: Ubiquity Press},
     title = {Greenland ice sheet. In: State of the Climate in 2021 - The Arctic},
     year = {2022},
     url = {https://doi.org/10.1175/BAMS-D-22-0082.1},
+    doi = {10.1175/BAMS-D-22-0082.1},
     volume = {103},
     issue = {8},	
     pages = {S276–S279},
@@ -182,6 +188,7 @@ Publisher: Ubiquity Press},
     year = {2023},
     version = {2},
     url = {https://doi.org/10.22008/FK2/VVXGUT},
+    doi = {10.22008/FK2/VVXGUT},
 }
 
 @misc{vandecrux-gcnet-2020,
