### 2023-03-07 155606e
Get module changed to fetch latest DOI using pyDataverse (#112)
* DOIs replaced with thredds links

* URL updater removed

* Bug fixes

* pyDataverse querying added (no token needed)

* urls lookup table retired

* look-up table implemented in module
