+++ src/pypromice/get.py
@@ -3,26 +3,20 @@
 """
 AWS data retrieval module
 """
+from pyDataverse.api import NativeApi
 import pandas as pd
+import xarray as xr
 import unittest, pkg_resources
 from datetime import datetime
-        
-def aws_names(url_index=None, delimiter=','):
+   
+def aws_names():
     '''Return PROMICE and GC-Net AWS names that can be used in get.aws_data() 
     fetching'''
-    if url_index is None:
-        with pkg_resources.resource_stream('pypromice', 'data_urls.csv') as stream:
-            lines = stream.read().decode("utf-8")
-            lines = lines.split("\n")
-    else:
-        stream = url_index 
-        with open(stream, 'r') as f:
-            lines = f.readlines()
-    names = [l.split(delimiter)[0] for l in lines]
-    print(f'Available dataset keywords: {names}')
-    return names    
-    
-def aws_data(aws_name):                                                        #TODO add daily and monthly datasets
+    lookup = lookup_table(['doi:10.22008/FK2/IW73UU', 'doi:10.22008/FK2/GNYFUK'])
+    print(f'Available dataset keywords: {list(lookup.keys())}')
+    return list(lookup.keys())
+
+def aws_data(aws_name):
     '''Return PROMICE and GC-Net AWS L3 v3 hourly observations
     
     Returns
@@ -30,135 +24,94 @@ def aws_data(aws_name):                                                        #
     df : pandas.DataFrame
         AWS observations dataframe
     '''
-    URL = _getURL(aws_name.lower()+'_hour')
-    df = pd.read_csv(URL, index_col=0, parse_dates=True)
-    return df
-    
-def watson_discharge_hourly():
+    lookup = lookup_table(['doi:10.22008/FK2/IW73UU', 'doi:10.22008/FK2/GNYFUK'])
+    assert aws_name.lower() in list(lookup.keys())
+    data = pd.read_csv(lookup[aws_name], index_col=0, parse_dates=True)
+    return data        
+  
+def watson_discharge(t='hour'):
     '''Return PROMICE hourly Watson river discharge
     
+    Parameters
+    ----------
+    t : str
+        Temporal resolution of the data - "hour", "day" or "year"
+
     Returns
     -------
     df : pandas.DataFrame
         Watson river discharge dataframe    
     '''
-    URL = _getURL('watson_discharge_hourly')
-    df = pd.read_csv(URL, sep="\s+", parse_dates=[[0,1,2,3]])\
-            .rename({"WaterFluxDiversOnly(m3/s)"         : "divers",
-                    "Uncertainty(m3/s)"                 : "divers_err",
-                    "WaterFluxDivers&Temperature(m3/s)" : "divers_t",
-                    "Uncertainty(m3/s).1"               : "divers_t_err",
-                    "WaterFluxCumulative(km3)"          : "Q",
-                    "Uncertainty(km3)"                  : "err"}, 
-                    axis='columns')
-    df = _getDFdatetime(df, list(df.iloc[:,0]))
+    lookup = lookup_table(['doi:10.22008/FK2/XEHYCM'])
+    
+    if 'year' in t.lower():
+        link = lookup['watson river discharge (1949-2021) yearly.txt']
+        df = pd.read_csv(link, sep="\s+", skiprows=9, index_col=0)
+        
+    elif 'daily' in t.lower() or 'day' in t.lower():
+        link = lookup['watson river discharge (2006-2021) daily.txt']   
+        df = pd.read_csv(link, sep="\s+", parse_dates=[[0,1,2]])\
+                        .rename({"WaterFluxDiversOnly(m3/s)"        : "divers",
+                                "Uncertainty(m3/s)"                 : "divers_err",
[…]
