+++ docs/requirements.txt
@@ -2,6 +2,7 @@ datetime
 eccodes
 numpy
 pandas
+pyDataverse
 scikit-learn
 sphinx_rtd_theme
 toml
