+++ MANIFEST.in
@@ -1,4 +1,3 @@
-include src/pypromice/*.csv
 include src/pypromice/process/*.csv
 include src/pypromice/tx/*.csv
 include src/pypromice/test/*
