### 2023-03-09 944b171
encode with i4 dtype (#113)
