### 2023-03-21 466030f
initial implementation of QC Static
