### 2023-03-22 8586a91
minor tweaks to compute_percentiles.py
