### 2023-03-23 #5298 @editorialbot
[REVIEW]: pypromice: A Python package for processing automated weather station data
**Submitting author:** @PennyHow (<a href="http://orcid.org/0000-0002-8088-8497">Penelope How</a>)
**Repository:** https://github.com/GEUS-Glaciology-and-Climate/pypromice
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.1
**Archive:** 10.22008/FK2/3TSBF0
