### 2023-03-24 0d9bece
fix season definitions in compute_percentiles.py
