### 2023-03-28 c766e9c
set percentile outliers to nan
