+++ src/pypromice/process/L1toL2.py
@@ -10,7 +10,7 @@ import os
 import xarray as xr
 import sqlite3
 
-from IPython import embed
+# from IPython import embed
 
 def toL2(L1, T_0=273.15, ews=1013.246, ei0=6.1071, eps_overcast=1., 
          eps_clear=9.36508e-6, emissivity=0.97):
@@ -152,6 +152,11 @@ def percentileQC(ds):
     ----------
     ds : xr.Dataset
         Level 1 dataset
+
+    Returns
+    -------
+    ds_out : xr.Dataset
+        Level 1 dataset with percentile outliers set to NaN
     '''
     stid=ds.station_id
     # Switch to pandas
@@ -159,6 +164,7 @@ def percentileQC(ds):
 
     # Define threshold dict to hold limit values, and 'hi' and 'lo' percentile.
     # Limit values indicate how far we will go beyond the hi and lo percentiles to flag outliers.
+    # *_u are used to calculate and define all limits, which are then applied to *_u, *_l and *_i
     var_threshold = {
         't_u': {'limit': 9}, # 'hi' and 'lo' held in separate 'seasons' dict
         'p_u': {'limit': 15},
@@ -166,13 +172,12 @@ def percentileQC(ds):
         'wspd_u': {'limit': 10}
         }
 
+    # Query from the sqlite db for specified percentiles
     con = sqlite3.connect('../qc/percentiles.db')
     cur = con.cursor()
-
-    # Query from the sqlite db for specified percentiles
-    # Different pattern for t_u, which considers seasons
     for k in var_threshold.keys():
         if k == 't_u':
+            # Different pattern for t_u, which considers seasons
             seasons = {1: {}, 2: {}, 3: {}, 4: {}}
             sql = f"SELECT p0p5,p99p5,season FROM {k} WHERE season in (1,2,3,4) and stid = ?"
             cur.execute(sql, [stid])
@@ -190,28 +195,58 @@ def percentileQC(ds):
 
     con.close() # close the database connection (and cursor)
 
-    # TO DO: Set outliers to NaN and write back to df
     for k in var_threshold.keys():
         if k == 't_u':
-            winter = df.t_u[df.index.month.isin([12,1,2])]
-            spring = df.t_u[df.index.month.isin([3,4,5])]
-            summer = df.t_u[df.index.month.isin([6,7,8])]
-            fall = df.t_u[df.index.month.isin([9,10,11])]
-            season_dfs = [winter,spring,summer,fall]
+            # use t_u thresholds to flag t_u, t_l, t_i
+            base_var = k.split('_')[0]
+            vars_all = [k, base_var+'_l', base_var+'_i']
+            for t in vars_all:
+                if t in df:
+                    winter = df[t][df.index.month.isin([12,1,2])]
+                    spring = df[t][df.index.month.isin([3,4,5])]
+                    summer = df[t][df.index.month.isin([6,7,8])]
+                    fall = df[t][df.index.month.isin([9,10,11])]
+                    season_dfs = [winter,spring,summer,fall]
 
-            for x1,x2 in zip([1,2,3,4], season_dfs):
-                lower_thresh = var_threshold[k]['seasons'][x1]['lo'] - var_threshold[k]['limit']
-                upper_thresh = var_threshold[k]['seasons'][x1]['hi'] + var_threshold[k]['limit']
+                    # _plot_percentiles_t(k,t,df,season_dfs,var_threshold,stid) # BEFORE OUTLIER REMOVAL
+                    for x1,x2 in zip([1,2,3,4], season_dfs):
+                        print(f'percentile flagging {t} {x1}')
[…]
