### 2023-03-29 5bdbfbc
clean up and comment L1toL2.py
