### 2023-03-29 5a17f96
add plotting capability to compute_percentiles.py
