+++ src/pypromice/qc/compute_percentiles.py
@@ -5,6 +5,7 @@ Compute percentile distributions and write to local sqlite db
 import sqlite3
 import os
 import pandas as pd
+import numpy as np
 from datetime import timedelta
 import argparse
 # from IPython import embed
@@ -23,6 +24,17 @@ def parse_arguments():
     return args
 
 def make_db_connection():
+    '''
+    Make connection to on-disk sqlite database
+
+    Parameters
+    ----------
+    None
+
+    Returns
+    -------
+    None
+    '''
     print('Creating sqlite3 connection...') 
     con = sqlite3.connect( # will create db if does not exist
     'percentiles.db', # define path and filename
@@ -34,6 +46,20 @@ def make_db_connection():
     return con, cur
 
 def create_tables(cur, var_list):
+    '''
+    Create variable-specific tables in the sqlite database
+
+    Parameters
+    ----------
+    cur : sqlite3.Cursor
+        cursor on the sqlite database connection
+    var_list : list
+        list of variable strings (e.g. 't_u') used to make percentiles
+
+    Returns
+    -------
+    None
+    '''
     for v in var_list:
         print(f'Creating {v} table...')
         if v not in ('t_u',):
@@ -72,12 +98,37 @@ def clear_tables(cur, var_list):
     '''
     Clear all rows from all tables. We run this by default, assuming that anytime
     we are running this script we intend to overwrite all rows for all tables.
+
+    Parameters
+    ----------
+    cur : sqlite3.Cursor
+        cursor on the sqlite database connection
+    var_list : list
+        list of variable strings (e.g. 't_u') used to make percentiles
+
+    Returns
+    -------
+    None
     '''
     for v in var_list:
         cur.execute(f'delete from {v}')
         print(f'Deleted {cur.rowcount} records from the {v} table.')
 
 def write_percentiles(cur, var_list):
+    '''
+    Write percentile data to tables
+
+    Parameters
+    ----------
+    cur : sqlite3.Cursor
+        cursor on the sqlite database connection
+    var_list : list
+        list of variable strings (e.g. 't_u') used to make percentiles
+
+    Returns
+    -------
+    None
+    '''
     print(f'writing to tables...')
     for x in os.walk(args.l3_filepath):
         if (len(x[2]) > 0): # files are present
@@ -132,6 +183,157 @@ def write_percentiles(cur, var_list):
                                 exe_list
                                 )
 
+def _analyze_percentiles():
+    '''
+    This is run ONLY to examine percentile thresholds with context to full datasets
+    '''
+    for x in os.walk(args.l3_filepath):
+        if (len(x[2]) > 0): # files are present
+            stid = x[0].split('/')[-1]
+            csv_file = [s for s in x[2] if '_hour.csv' in s]
+            if (len(csv_file) > 0) and (stid not in disclude_stations): # csv file is present
+                print(stid)
+                csv_filepath = x[0] + '/' + csv_file[0]
+                df = pd.read_csv(csv_filepath)
+                timestamp = pd.to_datetime(df.time)
+                df.set_index(timestamp, inplace=True)
+                _percentileQC(df, stid)
+
+def _percentileQC(df, stid):
+    '''
+    This is the same function that is found in L1toL2.py
+    Once thresholds are determined, they can be transferred to L1toL2.py
+    '''
+    # Define threshold dict to hold limit values, and 'hi' and 'lo' percentile.
[…]
