### 2023-03-30 260b439
Adapt examples to new package structure
- rename file to get with pget as the name before cannot be found
