### 2023-04-04 GEUS-Glaciology-and-Climate/pypromice#117 @jroettenbacher
[JOSS review] Documentation update
Dear pypromice Team,
I'm currently working on the review of your package. It's really great to see how easy it is to get your hands on data with this package! Great work!
I went through your documentation and tried the code examples in a local copy of the repository.

Some of the cells did not compute because of import errors which I figured is due to a change in package structure which hasn't found its way into the documentation yet. This is nothing major and can be easily fixed.

Since it is so easy to fix, I add a pull request to this issue which would fix the problem. I do not expect you to merge my pull request. This is just to show you the exact changes needed.

Best
Johannes
