### 2023-04-05 GEUS-Glaciology-and-Climate/pypromice#117 @PennyHow
Thanks very much. Ah yes, some of the documentation has not been updated since we did some restructuring. Thanks for the PR, we are happy to merge this and I will also check the rest of the documentation if there is anything else that needs updating.
