### 2023-04-05 GEUS-Glaciology-and-Climate/pypromice#117 @PennyHow
Documentation updated as part of #119. If you are happy with this then feel free to close this issue @jroettenbacher
