### 2023-04-05 #5298 @PennyHow
> I opened an [issue](https://github.com/GEUS-Glaciology-and-Climate/pypromice/issues/117) regarding a documentation update which would need to be fixed.

Thanks @jroettenbacher. Documentation has now been updated as part of the two most recent PRs.
