### 2023-04-17 GEUS-Glaciology-and-Climate/pypromice#135 @jroettenbacher
[JOSS review] Document tests 
Hi all,
as part of the JOSS review process I went through the repository and found some tests. However, those are not documented in the Documentation. I think this could fit in the [Developer Guide](https://pypromice.readthedocs.io/en/latest/guide_developer.html).

I see this as an enhancement to the usability but not absolutely necessary for acceptance.

Best
Johannes
