### 2023-04-17 GEUS-Glaciology-and-Climate/pypromice#136 @jroettenbacher
[JOSS review] Add plotting example to Documentation
Hi all,
as part of the JOSS review process I went through the Documenation of the package. I could imagine that including example plots of the data retrievable with the package can make it easier for users to understand the data and what can be achieved with it. Maybe this is out of scope for the Documentation since the package is rather technical. See it as a suggestion to improve user engagement and ease of access to the data.

This is not absolutely necessary for acceptance.

Best
Johannes
