### 2023-04-17 #5298 @jroettenbacher
Hi @observingClouds,
thanks for getting back to us. I am done with my review. The code works as expected and the documentation is sufficient to guide first time users and potential contributors to a point where the package is working and one can interact with data.

There are a few tests included in the package but those are not documented. I opened an [issue](https://github.com/GEUS-Glaciology-and-Climate/pypromice/issues/135) about this, but I don't see this as absolutely necessary for acceptance.
Another nice to have feature in the documentation would be some example plots of the data, so one can get a rough idea of what the data can be used for and how it could be plotted. I opened another [issue](https://github.com/GEUS-Glaciology-and-Climate/pypromice/issues/136) for that. Again, this is not absolutely necessary for acceptance.

Thus, I recommend the acceptance of the manuscript and package.

Best,
Johannes
