### 2023-04-19 GEUS-Glaciology-and-Climate/pypromice#135 @jroettenbacher
Indeed, these are the ones I meant. I'm not too familiar with testing in python so I didn't know about this command. I agree that adding this to the Developer Install section is a good idea.
