### 2023-04-20 a9025c0
Unit testing added to docs installation guide
