### 2023-04-20 076baf1
Plotting example shortened
