+++ docs/guide_user.rst
@@ -66,12 +66,12 @@ The Level 0 to Level 3 processing can also be executed from a CLI using the ``ge
     $ getL3 -c src/pypromice/test/test_config1.toml -i src/pypromice/test -o src/pypromice/test
 
 
-Loading PROMICE and GC-Net data
-===============================
+Loading our data
+================
 
 Import from Dataverse (no downloads!)
 -------------------------------------
-The automated weather station (AWS) datasets are openly available on our Dataverse_. These can be imported directly with pypromice, with no downloading required.
+The automated weather station (AWS) datasets from the PROMICE and GC-Net monitoring programmes are openly available on our Dataverse_. These can be imported directly with pypromice, with no downloading required.
 
 .. code:: python
 
@@ -136,8 +136,8 @@ If you would rather handle the AWS data as an ``xarray.Dataset`` object then the
 	ds = xr.Dataset.from_dataframe(df)
 
 
-Plotting PROMICE and GC-Net data
-================================
+Plotting our data
+=================
 
 Once loaded, variables from an AWS dataset can be simply plotted with using pandas or xarray.
 
@@ -156,7 +156,7 @@ Once loaded, variables from an AWS dataset can be simply plotted with using pand
 	airt = ds['t_u']
 	
 	# Plot with xarray
-	airt.plot()
+	ds['t_u'].plot()
 
 .. image:: https://raw.githubusercontent.com/GEUS-Glaciology-and-Climate/geus-glaciology-and-climate.github.io/master/assets/images/kpc_u_xr_plot.png
 
