+++ docs/guide_user.rst
@@ -151,19 +151,17 @@ Once loaded, variables from an AWS dataset can be simply plotted with using pand
 	
 .. code:: python
 	
-	# Extract variable for plotting
-	# In this case, we will extract air temperature
-	airt = ds['t_u']
-	
-	# Plot with xarray
+	# Plot variable with xarray
+	# In this case, we will plot air temperature
 	ds['t_u'].plot()
 
 .. image:: https://raw.githubusercontent.com/GEUS-Glaciology-and-Climate/geus-glaciology-and-climate.github.io/master/assets/images/kpc_u_xr_plot.png
 
 .. note::
 
-	For more complex plotting, please see either the xarray_ or pandas_ plotting documentation.
+	Variable names are provided in the dataset metadata, or can be found on in our variables look-up table here_. For more complex plotting, please see either the xarray_ or pandas_ plotting documentation.
 
+.. _here: https://github.com/GEUS-Glaciology-and-Climate/pypromice/blob/main/src/pypromice/process/variables.csv
 .. _xarray: https://docs.xarray.dev/en/stable/user-guide/plotting.html
 .. _pandas: https://pandas.pydata.org/docs/user_guide/10min.html#plotting
 	
