### 2023-04-21 #5298 @PennyHow
These comments have now been addressed in https://github.com/GEUS-Glaciology-and-Climate/pypromice/pull/137
