### 2023-04-22 08cd0c8
Documentation for unit testing and plotting (#137)
* Unit testing added to docs installation guide

* Plotting examples added

* Plotting example shortened

* Variable names info added

* PLocal package installation info added

* Ran to run
