### 2023-04-23 GEUS-Glaciology-and-Climate/pypromice#138 @simonrp84
[JOSS review] Command given to run tests is failing.
Hello,
In the [quick install guide](https://pypromice.readthedocs.io/en/latest/install.html) the command given to run the unittests (`python -m unittest discover pypromice`) throws an error:
```
(base) C:\Users\simon\pypromice>python -m unittest discover pypromice
Traceback (most recent call last):
  File "C:\Users\simon\miniconda3\lib\runpy.py", line 196, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "C:\Users\simon\miniconda3\lib\runpy.py", line 86, in _run_code
    exec(code, run_globals)
  File "C:\Users\simon\miniconda3\lib\unittest\__main__.py", line 18, in <module>
    main(module=None)
  File "C:\Users\simon\miniconda3\lib\unittest\main.py", line 100, in __init__
    self.parseArgs(argv)
  File "C:\Users\simon\miniconda3\lib\unittest\main.py", line 124, in parseArgs
    self._do_discovery(argv[2:])
  File "C:\Users\simon\miniconda3\lib\unittest\main.py", line 244, in _do_discovery
    self.createTests(from_discovery=True, Loader=Loader)
  File "C:\Users\simon\miniconda3\lib\unittest\main.py", line 154, in createTests
    self.test = loader.discover(self.start, self.pattern, self.top)
  File "C:\Users\simon\miniconda3\lib\unittest\loader.py", line 346, in discover
    raise ImportError('Start directory is not importable: %r' % start_dir)
ImportError: Start directory is not importable: 'pypromice'
```

Could you please investigate what's happening there? I tried a few different combinations and haven't had any luck working around the error.
