### 2023-05-01 ae336f0
Corrected missing air temp correction for z_stake
