### 2023-05-02 GEUS-Glaciology-and-Climate/pypromice#138 @simonrp84
Hi,
I can confirm that the commands you list here are working OK.
```
conda create --name pypromice-test python=3.8
conda activate pypromice-test
pip install --upgrade git+http://github.com/GEUS-Glaciology-and-Climate/pypromice.git
python -m unittest discover pypromice
```
But those are slightly different from what's written in the installguide, which are:
```
conda create --name pypromice python=3.8
conda activate pypromice
conda install xarray pandas pathlib
conda install -c conda-forge netCDF4
git clone git@github.com:GEUS-Glaciology-and-Climate/pypromice.git
cd pypromice/
pip install .
python -m unittest discover pypromice
```
These commands do not work unless I'm in the pypromice source directory. Perhaps you could edit the installguide to mirror your commands in the comment above? This may be some operating system specific issue, as I'm only able to test on Windows at the moment.
