### 2023-05-02 #5298 @simonrp84
I've now had a chance to look through this properly and overall it seems good!
I raised one issue about the tests, and have a few minor comments below:

- On github page: Add link to readthedocs as the website on the right hand upper side.

 - Would it be possible to create a conda version of the install to complement pip? Pip-only could be problematic for users as mixing pip and conda installs on the same environment can cause strange behaviour and issues.

- I think the index page for readthedocs could be improved. Can you add an additional sentence or two after "pypromice is designed for processing and handling PROMICE automated weather station (AWS) data." to describe what PROMICE data is and why a tool is needed to handle it? I know you mention this elsewhere, but it would be nice tohave that up-front in the docs.

- It's not immediately clear in the readthedocs pages where the example .toml files are located, it owuld be good to directly specify where the example toml files are from within the user guide section.

- This may be a lack of library knowledge on my part, but why does the 'developer install 'section of the docs say that netcdf, xarray, pandas and pathlib need to be installed via conda? All the other dependencies are installed via pip when running "pip install ."

- In the user guide, the code given in the Level 0 to Level 3 processing section is incorrectly indented, the leading space at the beginning of the lines should be removed.
