### 2023-05-03 GEUS-Glaciology-and-Climate/pypromice#138 @PennyHow
Yes, editing the installation guide is a good idea - saves confusion. We can do that no problem.
