### 2023-05-03 #5298 @PennyHow
Thanks for the comments @simonrp84!

> On github page: Add link to readthedocs as the website on the right hand upper side.

- [ ] Sure!

> Would it be possible to create a conda version of the install to complement pip? Pip-only could be problematic for users as mixing pip and conda installs on the same environment can cause strange behaviour and issues.

- [ ] I think this is a good point. We currently don't have a conda package for pypromice. Actually the pip install is only from the repo and not a pip package. We have been thinking about making a pip package, and a conda package would naturally be a good additional step. I would like to take a look at setting this up as an Action, so that conda and pip packages are generated automatically on repo releases. Let me take a look into this further and get back to you.

> I think the index page for readthedocs could be improved. Can you add an additional sentence or two after "pypromice is designed for processing and handling PROMICE automated weather station (AWS) data." to describe what PROMICE data is and why a tool is needed to handle it? I know you mention this elsewhere, but it would be nice tohave that up-front in the docs.

- [ ] Yes, absolutely. I've opened a [new branch](https://github.com/GEUS-Glaciology-and-Climate/pypromice/tree/joss-doc-edits) to address this modification to the documentation alongside [the issue you raised about the unit tests](https://github.com/GEUS-Glaciology-and-Climate/pypromice/issues/138).

> It's not immediately clear in the readthedocs pages where the example .toml files are located, it owuld be good to directly specify where the example toml files are from within the user guide section.

- [ ] Yes, I can add a section about using the example files to perform test runs.

> This may be a lack of library knowledge on my part, but why does the 'developer install 'section of the docs say that netcdf, xarray, pandas and pathlib need to be installed via conda? All the other dependencies are installed via pip when running "pip install ."

- [ ] I think here we were merely trying to demonstrate that conda and pip can be used together to create the pypromice install and package dependencies. However, I see that this may be confusing. We can change this to be clearer (and if we manage a conda install and a pip install then we can have instructions for both set-ups).

> In the user guide, the code given in the Level 0 to Level 3 processing section is incorrectly indented, the leading space at the beginning of the lines should be removed.

- [ ] Sure. This can be fixed.
