### 2023-05-04 fda442a
increment to v1.1.1 in setup.py
