### 2023-05-05 4d84d39
Acronym definition added
