### 2023-05-05 5c0903a
Index, installation and guides docs modified (#140)
* Docs updated for joss comments

* readthedocs build urllib3 workaround

* Acronym definition added
