### 2023-05-07 d693784
extrapolate gps_alt from entire record if needed
