### 2023-05-09 GEUS-Glaciology-and-Climate/pypromice#138 @simonrp84
Looks good to me, thanks for the rapid modifications!
