### 2023-05-09 #5298 @PennyHow
@simonrp84, a pypi package is now set up [here](https://pypi.org/project/pypromice/1.1.1/). This should update automatically when repo releases are made, as defined by the new Action on [this branch](https://github.com/GEUS-Glaciology-and-Climate/pypromice/tree/pypi-conda). Docs and readmes have also been updated on this branch.

I'll make a PR once we have come to an agreement about whether to include a conda package or not. I've been having a look at some way of making the conda package from the pypi package using `conda skeleton,` and therefore would be able to do this under the same Action triggered on a repo release. However, if it is not necessary for this review then I personally think we could work towards this in the future after some more stringent testing. It is still a very good suggestion and I am happy to continue working on it beyond this review. Let me know what you think and if this is okay @observingClouds.
