+++ src/pypromice/postprocess/wmo_config.py
@@ -13,11 +13,11 @@ https://confluence.ecmwf.int/display/ECC/WMO%3D32+element+table
 '''
 stid_to_skip = { # All the following IDS will not be processed or submitted
     'test': ['XXX'],
-    'not_registered': ['UWN'], # Need to register with Norwegion met
+    'not_registered': ['UWN','ZAK_A','WEG_L'], # Need to register UWN with Norwegion met
     'discontinued': ['CEN1','TAS_U','QAS_A','NUK_N','THU_U','JAR','SWC'],
     'no_instantaneous': ['ZAK_L','ZAK_U','KAN_B'], # currently not transmitting instantaneous values
     'suspect_data': [], # instantaneous data is suspect
-    'use_v3': ['KPC_L','KPC_U','NUK_U','ZAK_U','QAS_U'], # use v3 versions instead (but registered IDs are non-v3 names)
+    'use_v3': ['KPC_L','KPC_U','NUK_U','ZAK_L','ZAK_U','QAS_U'], # use v3 versions instead (but registered IDs are non-v3 names)
     'v3_bad': ['QAS_Lv3'] # QAS_Lv3 new ablation sensor & non-standard logger program. Must be addressed in field. Talk to RSF.
     # Eventually, as we change all stations to v3, the use_v3 list should be empty (there will not be v3 and non-v3 stations operating together)
     # If your remove a station from 'v3_bad', then add it to 'use_v3' if both v2 and v3 stations still exist together
