+++ src/pypromice/postprocess/csv2bufr.py
@@ -5,6 +5,7 @@ Post-processing functions for AWS station data, such as converting PROMICE and G
 """
 import pandas as pd
 import sys, traceback
+import os
 from datetime import datetime, timedelta
 from eccodes import codes_set, codes_write, codes_release, \
                     codes_bufr_new_from_samples, CodesInternalError, \
@@ -36,19 +37,27 @@ def getBUFR(s1, outBUFR, stid, land_stids):
         The station ID to be processed. e.g. 'KPC_U'
     land_stids : list
         List of station IDs for land-based stations
+
+    Returns
+    -------
+    remove_file : boolean
+        Status object to return to getBUFR indicating successful completion
     '''
+    remove_file = False
+
     # Open bufr file
     fout = open(outBUFR, 'wb')
 
-    # for i1, r1 in df1.iterrows(): # If dataframe passed w/ multiple rows
-
-    #Create new bufr message to write to
+    # Create new bufr message to write to
     ibufr = codes_bufr_new_from_samples('BUFR4')
     timestamp = datetime.strptime(s1['time'], '%Y-%m-%d %H:%M:%S')
     config_key = 'mobile'
     if stid in land_stids:
         config_key = 'land'
     try:
+        # we must pass all the following functions without error.
+        # If handled (or unhandled) errors occur, we re-raise and
+        # the exceptions below will set remove_file to True.
         setTemplate(ibufr, timestamp, stid, config_key)
         setStation(ibufr, stid, config_key)
         setAWSvariables(ibufr, s1, timestamp, stid)
@@ -61,17 +70,24 @@ def getBUFR(s1, outBUFR, stid, land_stids):
 
     except CodesInternalError as ec:
         print(traceback.format_exc())
-        print(ec)
-        sys.exit('-----> CodesInternalError in getBUFR!')
+        # print(ec)
+        print(f'-----> CodesInternalError in getBUFR for {stid}!')
+        remove_file = True
     except Exception as e:
         # Catch anything else here...
         print(traceback.format_exc())
-        print(e)
-        sys.exit('!!!!!!!!!! ERROR in getBUFR')
+        # print(e)
+        print(f'-----> ERROR in getBUFR for {stid}')
+        remove_file = True
 
     codes_release(ibufr)
 
     fout.close()
+
+    if remove_file is True:
+        print(f'-----> Removing file for {stid}')
+        os.remove(fout.name)
+    return remove_file
 
 
 def setTemplate(ibufr, timestamp, stid, config_key):
@@ -129,15 +145,16 @@ def setStation(ibufr, stid, config_key):
                 stid = stid.replace('_O','')
             if stid == 'CEN2':
                 stid = 'CEN'
-            if stid in v:
+            try:
                 codes_set(ibufr, k, v[stid])
-            else:
-                sys.exit('!!!!!!!!!! ID not found for {}'.format(stid))
+            except KeyError as e:
+                print(f'-----> ID not found for {stid}')
+                raise # throw error back to getBUFR where it is handled
         else:
             if codes_is_defined(ibufr, k) == 1:
                 codes_set(ibufr, k, v)
             else:
-               print('-----> setStation Key not defined: {}'.format(k))
+               print(f'-----> setStation Key for {stid} not defined: {k}')
                continue
 
 
@@ -222,12 +239,13 @@ def setBUFRvalue(ibufr, b_name, value):
             codes_set(ibufr, b_name, value)
         except CodesInternalError as ec:
             print(f'{ec}: {b_name}')
-            sys.exit('-----> CodesInternalError in setBUFRvalue!')
+            print('-----> CodesInternalError in setBUFRvalue!')
+            raise # throw error back to getBUFR where it is handled
[…]
