### 2023-05-12 #5298 @PennyHow
Hi again @observingClouds, I've been testing a conda build of pypromice and all is working okay. The only thing is that one of our dependencies, `pydataverse`, is not available as a conda package and all dependencies must be conda builds in order for it to work.

So the options are to either
- include `pydataverse` as an optional dependency in the conda build, so therefore users have to install it themselves
- pursue getting a conda build for `pydataverse`

This is quite an important dependency as it is how our users can access our datasets through the `pypromice.get` module. Therefore I'm currently trying to implement the latter option and have [opened an issue with pydataverse](https://github.com/gdcc/pyDataverse/issues/159). I've successfully made a conda build for `pydataverse` myself, so I am hoping they will see that it is relatively straightforward.

Again, I am happy to continue pursuing this outside this JOSS review and hope this demonstrates that we have the ball rolling on getting pypromice a conda distribution.
