### 2023-05-15 #5298 @observingClouds
Hi @PennyHow,
Please excuse the delay. I totally agree that the dependency should not further delay the acceptance of this submission.
A few items are left to do now before I can recommend the acceptance. Could you please:

- [ ] Make a tagged release of your software, and list the version tag of the archived version here. You might want to consider merging the pip-branch beforehand.

Thank you!
