### 2023-05-16 f51c4d7
pypi build for JOSS review (#143)
* pypi publish action added

* pypi token added

* pip install updated
