+++ docs/index.rst
@@ -9,7 +9,7 @@ If you intend to use PROMICE and GC-Net AWS data and/or pypromice in your work, 
 
 Fausto, R.S., van As, D., Mankoff, K.D., Vandecrux, B., Citterio, M., Ahlstrøm, A.P., Andersen, S.B., Colgan, W., Karlsson, N.B., Kjeldsen, K.K., Korsgaard, N.J., Larsen, S.H., Nielsen, S., Pedersen, A.Ø., Shields, C.L., Solgaard, A.M., and Box, J.E. (2021) Programme for Monitoring of the Greenland Ice Sheet (PROMICE) automatic weather station data, Earth Syst. Sci. Data, 13, 3819–3845, https://doi.org/10.5194/essd-13-3819-2021
 
-How, P., Wright, P.J., Mankoff, K. and Vandecrux, B. (2023) pypromice v1.0.0, GEUS Dataverse, https://doi.org/10.22008/FK2/3TSBF0
+How, P., Wright, P.J., Mankoff, K., Vandecrux, B., Fausto, R.S. and Ahlstrøm, A.P. (2023) pypromice, GEUS Dataverse, https://doi.org/10.22008/FK2/3TSBF0
 
 .. _pypromice: https://github.com/GEUS-Glaciology-and-Climate/pypromice
 .. _PROMICE: https://promice.dk
