### 2023-05-24 #5298 @PennyHow
Hi again @observingClouds, I've just been told that the DOIs for our weather station dataset are being merged into one, so they will have a different DOI to those cited in the JOSS paper. Would it be possible to edit the paper with this update?
