### 2023-05-24 #5298 @observingClouds
Hi @PennyHow,
As DOIs should be persistent identifiers they should continue to point to the same datasets, but if you like to cite a collection DOI or a newer version of the dataset, please go ahead and change the paper and references accordingly and we'll see what we can do.
