+++ paper/paper.md
@@ -73,7 +73,7 @@ The `pypromice` software handles data from 43 AWSs on hourly, daily and monthly 
 
 GEUS is responsible for the Programme for Monitoring of the Greenland Ice Sheet ([PROMICE](https://promice.org)), which is now a network of over 21 AWSs installed across the Greenland Ice Sheet [@ahlstrom-programme-2008]. Launched in 2007, these one-level tripod stations are designed to stand on ice and move with the ice flow close to the ice sheet periphery [@fausto-programme-2021; @how-one-boom-2022]. The PROMICE stations are designed to monitor the surface melt and its meteorological drivers in the ablation area of the ice sheet.
 
-In 2021, GEUS assumed responsibility of the Greenland Climate Network (GC-Net) AWS locations [@steffen-greenland-1996], previously maintained by the United States National Science Foundation (NSF), National Aeronautics and Space Administration (NASA) and Swiss Federal Institute for Forest, Snow and Landscape Research (WSL). This expansion added 16 two-level mast stations to GEUS' sites. The data from these stations are intended to monitor conditions on the inner regions of the ice sheet, including snow accumulation and surface conditions [@how-two-boom-2022].
+In 2021, GEUS assumed responsibility of the Greenland Climate Network (GC-Net) AWS locations [@steffen-greenland-1996], previously maintained by the United States National Science Foundation (NSF), National Aeronautics and Space Administration (NASA) and Swiss Federal Institute for Forest, Snow and Landscape Research (WSL). This expansion added 16 two-level mast stations to GEUS' sites. The data from these stations are intended to monitor conditions on the inner regions of the ice sheet, including snow accumulation and surface conditions [@how-one-boom-2022].
 
 The Greenland Ecosystem Monitoring programme ([GEM](https://g-e-m.dk)) is an integrated, long-term monitoring effort that examines the effects of climate change on Arctic ecosystems. Established in 1995, GEM includes monitoring at Zackenberg, Kobbefjord, and Disko, Greenland. The program offers access to over 1000 freely-available environmental datasets, including data from 6 GEUS-designed AWS installations [@gem-glaciobasis-2020] which have been used in scientific publications [@messerli-snow-2022].
 
