+++ paper/paper.bib
@@ -69,8 +69,8 @@
 	url = {https://doi.org/10.22008/FK2/3TSBF0},
 	urldate = {2023-05-22},
 	publisher = {GEUS Dataverse},
-	author = {How, Penelope and Wright, Patrick J. and Mankoff, Kenneth D. and Vandecrux, Baptiste and Fausto, Robert S. and Ahlstrøm, Andreas P.},
-	year = {2023},        month = may,
+	author = {How, P. and Wright, P. J. and Mankoff, K. D. and Vandecrux, B. and Fausto, R. S. and Ahlstrøm, A. P.},
+	year = {2023},
 	doi = {10.22008/FK2/IPOHT5},
     	version = {1.2.0},
 }
