+++ paper/paper.md
@@ -64,7 +64,7 @@ A defined processing workflow is included in `pypromice` for transforming origin
 3. Supply easy and accessible methods to handle AWS data
 4. Provide opportunities to contribute to the processing and handling of AWS data in an open and collaborative manner
 
-The `pypromice` software has been designed to handle and process data from AWSs located in Greenland. The compilation and processing of data from national AWS networks has historically been conducted through un-distributed, oftentimes proprietary software. Similar Python packages to `pypromice` has been developed to process data from historical AWS in Greenland [@vandecrux-gcnet-2020;@steffen-gcnet-2023] or commercial AWS (e.g. [pywws](https://pypi.org/project/pywws/) [@easterbrook-pywws-2023]) or to post-process and harmonize AWS data from different institutions (e.g. [JAWS](https://github.com/jaws/jaws) [@zender-jaws-2019]). Therefore, there was a key need for the development of `pypromice` in order to have a package with a complete and operational `L0` to `L3` workflow.
+The `pypromice` software has been designed to handle and process data from AWSs located in Greenland. The compilation and processing of data from national AWS networks has historically been conducted through un-distributed, oftentimes proprietary software. Similar Python packages to `pypromice` have been developed to process data from historical AWS in Greenland [@vandecrux-gcnet-2020;@steffen-gcnet-2023], from commercial AWS (e.g. [pywws](https://pypi.org/project/pywws/) [@easterbrook-pywws-2023]), or to post-process and harmonize AWS data from different institutions (e.g. [JAWS](https://github.com/jaws/jaws) [@zender-jaws-2019]). Therefore, there was a key need for the development of `pypromice` in order to have a package with a complete and operational `L0` to `L3` workflow.
 
 
 # Usage
@@ -80,7 +80,7 @@ The Greenland Ecosystem Monitoring programme ([GEM](https://g-e-m.dk)) is an int
  
 # Documentation
 
-`pypromice` versions accompany releases of GEUS AWS one-boom and two-boom data publications [@how-pypromice-2022].
+`pypromice` versions accompany releases of GEUS AWS data publications [@how-pypromice-2022].
 
 Package documentation is available on the `pypromice` [readthedocs](https://pypromice.readthedocs.io/en/latest/). 
 
