### 2023-05-25 625e510
DOIs updated in JOSS paper (#147)
* dataset DOI updated

* Citations updated

* authorship changed

* two-boom mentions removed
