+++ paper/paper.bib
@@ -56,34 +56,23 @@
 }
 
 @article{how-one-boom-2022,
-	title = {AWS one boom tripod Edition 4},
+	title = {PROMICE and GC-Net automated weather station data in Greenland},
 	url = {https://doi.org/10.22008/FK2/IW73UU},
 	journal = {Dataset},
-	author = {How, Penelope and Mankoff, Kenneth D. and Wright, Patrick J. and Vandecrux, Baptiste and Ahlstrøm, Andreas P. and Fausto, Robert S.},
-	month = jan,
+	author = {How, P. and Abermann, J. and Ahlstrøm, A.P. and Andersen, S.B. and Box, J. E. and Citterio, M. and Colgan, W. T. and Fausto, R. S. and Karlsson, N. B. and Jakobsen, J. and Langley, K. and Larsen, S. H. and Mankoff, K. D. and Pedersen, A. Ø. and Rutishauser, A. and Shield, C. L. and Solgaard, A. M. and van As, D. and Vandecrux, B. and Wright, P. J.},
 	year = {2023},
 	doi = {10.22008/FK2/IW73UU},
-}
-
-@article{how-two-boom-2022,
-	title = {AWS two boom mast Edition 1},
-	url = {https://doi.org/10.22008/FK2/GNYFUK},
-	journal = {Dataset},
-	author = {How, Penelope and Mankoff, Kenneth D. and Wright, Patrick J. and Vandecrux, Baptiste and Ahlstrøm, Andreas P. and Fausto, Robert S.},
-	month = jan,
-	year = {2023},
-	doi = {10.22008/FK2/GNYFUK},
 }
 
 @misc{how-pypromice-2022,
 	title = {pypromice},
 	url = {https://doi.org/10.22008/FK2/3TSBF0},
-	urldate = {2023-01-16},
+	urldate = {2023-05-22},
 	publisher = {GEUS Dataverse},
-	author = {How, Penelope and Wright, Patrick J. and Mankoff, Kenneth D. and Vandecrux, Baptiste},
+	author = {How, P. and Wright, P. J. and Mankoff, K. D. and Vandecrux, B. and Fausto, R. S. and Ahlstrøm, A. P.},
 	year = {2023},
 	doi = {10.22008/FK2/IPOHT5},
-    	version = {1.0.0},
+    	version = {1.2.0},
 }
 
 @article{hoyer-xarray-2017,
