### 2023-05-25 GEUS-Glaciology-and-Climate/pypromice#148 @PennyHow
pypromice conda distribution
A comment from the JOSS review was whether `pypromice` could be made available as a conda package, alongside the already-available pypi package. We could not complete this task in the JOSS review as one of our dependencies did not have a conda distiribution, and conda packages require that all dependencies are also available as conda packages (this is so the build can go as smoothly as possible). You can see the full explanation [here in the JOSS review](https://github.com/openjournals/joss-reviews/issues/5298#issuecomment-1545554136).

The package in question was [pyDataverse](https://github.com/gdcc/pyDataverse) which we use in pypromice to fetch the latest releases of our datasets. It only had pypi package distribution previously and not conda. Over the course of the last week or so, I have made and published a conda-forge release with the pyDataverse team - the issue and discussion can be found [here](https://github.com/gdcc/pyDataverse/issues/159), alongside the PR which is [here](https://github.com/gdcc/pyDataverse/pull/160).

Now we have [a conda-forge distribution of pyDataverse](https://anaconda.org/conda-forge/pydataverse), we can move on and get pypromice a conda distribution too. This involves forking the conda-forge recipes repo and building a pypromice feedstock. I can do this fairly quickly. I just need confirmation from the chief developers @mankoff @BaptisteVandecrux @patrickjwright that you are happy to be listed as maintainers of the conda distribution? - by doing so, you also become part of the conda-forge organization on Github which is pretty cool.
