### 2023-05-25 #5298 @observingClouds
Hi @PennyHow,
Just a quick clarification question. Is it intentional that the new common dataset has the same DOI (https://doi.org/10.22008/FK2/IW73UU) as the single dataset beforehand?
