### 2023-05-25 #5298 @PennyHow
Yes it is intentional. Previously we had two DOIs for our weather data, differentiated by the type of weather station - our one-boom tripod design or our two-boom mast design.

However, some of my colleagues were recently at EGU where we had some feedback from our users that they would rather have all weather data in one place (i.e. under the same DOI). As long as the station type is provided in the dataset readme and the metadata of the data file (which it is) then we did not see a problem with this.

So the intention was to merge everything under one DOI. As we already have two pre-existing DOIs for our weather data, it made sense to retire one of these (it still exists, just no longer being updated) and migrate everything to the other active one.

I didn't realise that my colleagues wanted this implemented asap, hence why this has happened now. And I felt it best to ask for it to be changed in this paper otherwise it would already be out-dated before publication.
