### 2023-05-25 #5298 @observingClouds
Thanks for the clarification @PennyHow. I would generally prefer the creation of a new DOI for an updated dataset, but because this submission only mentions the dataset for general purposes and no specific version of it, this should be fine and maybe advantageous as it always points to the landing page containing all versions of the dataset. If someone would use a specific version of the dataset to derive a new product or conducts an analysis, the DOI should point to an unambiguous dataset, i.e. a specific version.
