### 2023-05-26 GEUS-Glaciology-and-Climate/pypromice#148 @PennyHow
> No need to put me at this point, since I don't really know what it is to maintain a conda distribution. Thanks for asking!

Okay no worries. It's not too complicated stuff to handle to be a maintainer, but I understand. You can always be added in later if you change your mind!
