### 2023-05-30 #5298 @kthyng
@PennyHow I see that the version at your archive is v1.2.1 — should the version associated with your JOSS paper (this submission) also be v1.2.1?

Also we prefer if possible for the title and author list at your software archive to exactly match your JOSS paper if you're willing.
