### 2023-06-01 #5298 @PennyHow
> * Please make sure all references are correctly inline vs. parenthetical. For example, the last reference on page 1 should be inline instead of parenthetical, so you should remove the [] around it.

I've checked this and altered references which should be inline. These changes, along with the others listed here, have been made in our most recent PR https://github.com/GEUS-Glaciology-and-Climate/pypromice/pull/150

> * why is there an extra line in the list on page 3?

There was a stray tab at the end of the first bulletpoint. This has now been removed.

> * Also please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

Done. I have altered references where capitalization should be, e.g. placenames, project names.

> @PennyHow I see that the version at your archive is v1.2.1 — should the version associated with your JOSS paper (this submission) also be v1.2.1?

Yes, we just has a minor patch to fix. So feel free to change the associated version to v1.2.1.

> Also we prefer if possible for the title and author list at your software archive to exactly match your JOSS paper if you're willing.

We would rather have our software archive reflect only the title of the Python package "pypromice" as this is a rolling archive that continues to update as the software develops. Would a good compromise be that we refer to the JOSS paper in the description/metadata of the software archive instead?
