### 2023-06-01 e00e92f
JOSS paper typesetting edits (#150)
* reference capitalization

* extra line removed

* reference parentheses changed

* Modification to pre-existing package references
