+++ paper/paper.bib
@@ -1,7 +1,7 @@
 
 @article{ahlstrom-programme-2008,
     author = {Ahlstrøm, A. P. and {the PROMICE project team}},
-    title = {A new programme for monitoring the mass loss of the Greenland ice sheet},
+    title = {A new programme for monitoring the mass loss of the {Greenland} {Ice} {Sheet}},
     year = {2008},
     url = {https://doi.org/10.34194/geusb.v15.5045},
     doi = {10.34194/geusb.v15.5045},
@@ -12,7 +12,7 @@
 
 @article{box-greenland-2022,
     author = {Box, J.E. and Hubbard, A. and Bahr, D. B. and Colgan, W. T. and Fettweis, X. and Mankoff, K. D. and Wehrlé, A. and Noël, B. and van den Broeke, M. R. and Wouters, B. and Bjørk, A. A. and Fausto R. S.},
-    title = {Greenland ice sheet climate disequilibrium and committed sea-level rise},
+    title = {{Greenland} {Ice} {Sheet} climate disequilibrium and committed sea-level rise},
     year = {2022},
     doi = {10.1038/s41558-022-01441-2},
     url = {https://doi.org/10.1038/s41558-022-01441-2},
@@ -23,14 +23,14 @@
 
 @misc{easterbrook-pywws-2023,
     author = {Easterbrook, Jim },
-    title = {Python software for USB Wireless Weather Stations},
+    title = {Python software for {USB} {Wireless} {Weather} {Stations}},
     year = {2008 - 2023},
     version = {23.2.0},
     url = {https://github.com/jim-easterbrook/pywws},
 }
 
 @article{fausto-programme-2021,
-	title = {Programme for Monitoring of the Greenland Ice Sheet (PROMICE) automatic weather station data},
+	title = {{Programme} for {Monitoring} of the {Greenland} {Ice} {Sheet} ({PROMICE}) automatic weather station data},
 	volume = {13},
 	issn = {1866-3508},
 	url = {https://essd.copernicus.org/articles/13/3819/2021/},
@@ -46,7 +46,7 @@
 }
 
 @article{gem-glaciobasis-2020,
-	title = {GlacioBasis Zackenberg - Snow cover - Snow depth radar},
+	title = {{GlacioBasis} {Zackenberg} - Snow cover - Snow depth radar},
 	url = {https://doi.org/10.17897/E594-NV64},
 	doi = {10.17897/E594-NV64},
 	urldate = {2022-12-15},
@@ -56,27 +56,26 @@
 }
 
 @article{how-one-boom-2022,
-	title = {PROMICE and GC-Net automated weather station data in Greenland},
+	title = {{PROMICE} and {GC}-{Net} automated weather station data in {Greenland}},
 	url = {https://doi.org/10.22008/FK2/IW73UU},
-	journal = {Dataset},
-	author = {How, P. and Abermann, J. and Ahlstrøm, A.P. and Andersen, S.B. and Box, J. E. and Citterio, M. and Colgan, W. T. and Fausto, R. S. and Karlsson, N. B. and Jakobsen, J. and Langley, K. and Larsen, S. H. and Mankoff, K. D. and Pedersen, A. Ø. and Rutishauser, A. and Shield, C. L. and Solgaard, A. M. and van As, D. and Vandecrux, B. and Wright, P. J.},
+	publisher = {GEUS Dataverse},
[…]
