### 2023-06-01 #5298 @kthyng
> We would rather have our software archive reflect only the title of the Python package "pypromice" as this is a rolling archive that continues to update as the software develops. Would a good compromise be that we refer to the JOSS paper in the description/metadata of the software archive instead?

Ok no biggie. That sounds like a reasonable compromise, or you can just skip it.

Looks like we are ready to go!
