### 2023-06-02 #5298 @PennyHow
Thanks very much @kthyng! This was a great publishing experience and we are very excited to have our work with JOSS now. And thanks again @observingClouds @simonrp84 @jroettenbacher!
